requirements.txt
cohere
pyttsx3
SpeechRecognition
wikipedia-api
opencv-python
requests
numpy
ultralytics
